\name{dbt.Distances-package}
\alias{dbt.Distances-package}
\alias{dbt.Distances}
\docType{package}
\title{
Package containing functions for distance calculations
}
\description{
Functions for distance calculations using different metrics.  
}
\details{
\tabular{ll}{
Package: \tab dbt.Distances\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-02-22\cr
License: \tab  \cr
LazyLoad: \tab yes\cr
}
}
\author{
Raphael Paebst
Maintainer: <ultsch@mathematik.uni-marburg.de>
}
\keyword{ package }