\name{Dist-package}
\Rdversion{1.1}
\alias{Dist-package}
\alias{Dist}
\docType{package}
\title{
Distance functions

}
\description{
Package contains different functions for calculations of distances
}
\details{
\tabular{ll}{
Package: \tab dbt.Dist\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-10-19\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Michael Thrun
}

\keyword{ package }
