% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{distance_columns}
\alias{distance_columns}
\title{Distance matrix columns}
\usage{
distance_columns(distances, column_indices, row_indices = NULL)
}
\arguments{
\item{distances}{A \code{\link{distances}} object.}

\item{column_indices}{An integer vector with point indices indicating
which columns to be extracted.}

\item{row_indices}{If \code{NULL}, complete rows will be extracted.
If integer vector with point indices, only the indicated
rows will be extracted.}
}
\value{
Returns a matrix with the requested columns.
}
\description{
\code{distance_columns} extracts columns from the distance matrix.
}
\details{
If the complete distance matrix is desired, \code{\link{distance_matrix}} is
faster than \code{distance_columns}.
}
