% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raw.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Distance matrix}
\usage{
distance_matrix(distances, point_indices = NULL)
}
\arguments{
\item{distances}{...}

\item{point_indices}{...}
}
\value{
...
}
\description{
Distance matrix
}
