% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raw.R
\name{distance_rows}
\alias{distance_rows}
\title{Distance rows}
\usage{
distance_rows(distances, row_indices, column_indices = NULL)
}
\arguments{
\item{distances}{...}

\item{row_indices}{...}

\item{column_indices}{...}
}
\value{
...
}
\description{
Distance rows
}
