% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances-package.R
\docType{package}
\name{distances-package}
\alias{distances-package}
\alias{distances}
\title{distances: Distance metric tools for R}
\description{
Tools for distances and metrics.
}
\details{
This package is under development, please exercise caution when using it.

More information and the latest version is found here:
\url{https://github.com/fsavje/distances}.

Bug reports and suggestions are greatly appreciated. They
are best reported here:
\url{https://github.com/fsavje/distances/issues}.
}
