\name{EuclidDist2all}
\alias{EuclidDist2all}
\title{ Euclidic distance of x to all rows in Data }
\arguments{
\item{x(1:d)}{row vector of length d}
\item{Data(1:n,1:d)}{matrix of size nxd}
}
\value{DistToAll(1:n)   column vector with distances from x to all rows in Data}
\description{
 Euclidic distance of x to all rows in Data }
\author{Michael Thrun
}
