\name{fastPdist}
\alias{fastPdist}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
fastPdist
}
\description{
calculates pairwise euclidean distances
}
\usage{
fastPdist(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
[1:n,1:m]      data to calculate distances to
}
}

\value{
dist[1:n,1:n]   distances
}

\author{
Michael Thrun
}

\keyword{fastPdist}% use one of  
