\name{InnerVariableDifferences}
\alias{InnerVariableDifferences}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
InnerVariableDifferences
}
\description{
# all inner differences di = xi-yi for all xi,yi in Variable
}
\usage{
InnerVariableDifferences(Variable)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Variable}{
(1:n)    matrix of size nxd without NaNs in columns where defined = 1

}
}

\value{
Di     matrix of all inner differences
}

\author{
MT
}