\name{jaccard}
\alias{jaccard}
\title{Computes dissimilarity indices Jaccard }
\usage{
jaccard(X)
}
\arguments{
\item{X}{Distance Matrix}
}
\value{ Kosinusdistanz der beiden Vektoren x,y}
\description{
The function computes dissimilarity indices Jaccard, which index is computed as 2B/(1+B), where B is Bray-Curtis dissimilarity }
\author{
MT
}