% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{max_distance_search}
\alias{max_distance_search}
\title{Max distance search}
\usage{
max_distance_search(distances, query_indices = NULL, search_indices = NULL)
}
\arguments{
\item{distances}{...}

\item{query_indices}{...}

\item{search_indices}{...}
}
\value{
...
}
\description{
Max distance search
}
