% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{nearest_neighbor_search}
\alias{nearest_neighbor_search}
\title{Nearest neighbor search}
\usage{
nearest_neighbor_search(distances, k, query_indices = NULL,
  search_indices = NULL, radius = NULL)
}
\arguments{
\item{distances}{...}

\item{k}{...}

\item{query_indices}{...}

\item{search_indices}{...}

\item{radius}{...}
}
\value{
...
}
\description{
Nearest neighbor search
}
