\name{pdistToroid}
\alias{pdistToroid}
\title{  Calculate toroid Euclidean Distances }
\usage{
pdistToroid(X,Y,Lines,Columns)
}
\arguments{
\item{X}{ vector of grid koordinated on a planar grid of size Lines}
\item{Y}{ vector of grid koordinated on a planar grid of size Columns}
\item{Lines}{lines of the planar grid }
\item{Columns}{columns of the planar grid }
}
\value{distance(s) between row(s) of x and y in MATHLBs pdist form}
\description{
  Calculate toroid Euclidean Distances }
\author{
MT}
