\name{squareform}
\alias{squareform}
\title{Format or generate a distance matrix. }
\usage{squareform(X)}
\arguments{
\item{X}{numeric vector or matrix}
}
\value{y      Returns a matrix if x is a vector, and a vector if x is a matrix.}
\description{
 Format or generate a distance matrix. 
 uses pracmas squareform with small changes
 }
\author{
Michael Thrun
}
