\name{supportRadiusVoting}
\alias{supportRadiusVoting}
\title{ Support Radius Voting}
\description{Calculates the support radius for all data points.}
\usage{
supportRadiusVoting(data)
}
\arguments{
\item{data}{A matrix containing the data points as rows }
}
\value{
\item{individualSupportRadius}{A vector containing the individual support radius.}
}
\author{ Raphael Paebst}