\name{VariablePrecision}
\alias{VariablePrecision}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
VariablePrecision
}
\description{
find he smallest (NonZero) Difference of any Value in Variable
}
\usage{
VariablePrecision(Variable)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Variable}{
(1:n,1:d)    numeric data or
}
}
\value{
 MinAbsDiff(1:d)        Absolute value of the smallest Difference of any Value in Variable
 
 MinAbsNZDiff(1:d)      Absolute value of the smallest NonZero Difference of any Value in Variable
 
 MinExpo(1:d)           The smalest Nachkommastelle in Variable NOT JET IMPLEMENTED

}

\author{
MT
}
