% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Distance matrix}
\usage{
distance_matrix(distances, indices = NULL)
}
\arguments{
\item{distances}{A \code{\link{distances}} object.}

\item{indices}{If \code{NULL}, the complete distance matrix is made.
If integer vector with point indices,
a partial matrix including only the indicated data points is made.}
}
\value{
Returns a distance matrix of class \code{\link[stats]{dist}}.
}
\description{
\code{distance_matrix} makes distance matrices (complete and partial) from
\code{\link{distances}} objects.
}
