% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances-package.R
\docType{package}
\name{distances-package}
\alias{distances-package}
\title{distances: Tools for Distance Metrics}
\description{
The \code{distances} package provides tools for constructing, manipulating
and using distance metrics in R. It calculates distances only as needed
(unlike the standard \code{\link[stats]{dist}} function which derives the
complete distance matrix when called). This saves memory and can increase
speed. The package also includes functions for fast nearest and farthest
neighbor searching.
}
\details{
See the package's website for more information:
\url{https://github.com/fsavje/distances}.

Bug reports and suggestions are greatly appreciated. They
are best reported here:
\url{https://github.com/fsavje/distances/issues/new}.
}
