% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.distances}
\alias{is.distances}
\title{Check \code{distances} object}
\usage{
is.distances(x)
}
\arguments{
\item{x}{object to check.}
}
\value{
Returns \code{TRUE} if \code{x} is a valid
        \code{distances} object, otherwise \code{FALSE}.
}
\description{
\code{is.distances} checks whether the provided object
is a valid instance of the \code{\link{distances}} class.
}
